+++ packages/fast-components/src/progress/progress-ring/progress-ring.stories.ts
@@ -1,0 +1,13 @@
+import { FASTProgressRing } from ".";
+import { FASTDesignSystemProvider } from "../../design-system-provider";
+import Examples from "./fixtures/circular.html";
+
+// Prevent tree-shaking
+FASTProgressRing;
+FASTDesignSystemProvider;
+
+export default {
+    title: "Progress Ring",
+};
+
+export const Base = () => Examples;